move mock implementaions to the replica file
